cleaned up insertion script and commentary.
