+++ src/data_pipeline/run_unverified_test_center_diff_merges.js
@@ -1,26 +1,22 @@
-/* 
-TODO: some refactoring would be good:
-1. maybe change the diff file methodology. Consider storing more contextual data like commit date, and a commit message.
+/*
+This script is intended to be run on startup of our APIDockerfile. It's called from the dockerfile's ENTRYPOINT,
+via entrypoint.sh.
 
-2. consider moving the code to a data_pipeline folder in /server, merged with Python pipeline code. All data pipeline code in the same directory.
+The script is comparable to a database seed operation, and seeds the UnverifiedTestCenters table with new data. 
+The data originally comes from a variety of sources, is dumped to CSV files, and merged to the system using the cmd_load_new_test_centers.py tool.
 
-3. move helper functions into separate files, e.g. AWS, file utilities, DB utilities
+The seed process in this script does the following:
+1. It loads the CHRONOLOGICAL_DIFF_KEYS JSON object
+2. It checks against the UnverDiff database table. If any of the keys aren't present in the database, those keys will be processed, in order.
+3. To process a key, a 'diff' file is downloaded from S3. The file contains prepared objects to be inserted into the UnverifiedTestCenters table.
+4. The objects are inserted into the database. This process repeats for each new key. If any key fails, the process terminates and the deployment fails.
 
-4. clean up console logging, reduce some of the excess (but keep the logging, it will be helpful for debugging)
+Deployment failure due to this process should be fatal. The quick fix is to remove the aggrevating key from the CHRONOLOGICAL_DIFF_KEYS file.
 
 */
 
-const path = require('path');
-const MODELS_DIR = path.join(__dirname, '../server/db', 'models')
-const db = require(MODELS_DIR)
-const { Op } = require('sequelize')
-const fs = require('fs').promises;
-const AWS = require('aws-sdk');
-const s3 = new AWS.S3();
-
-const S3_UNVER_DIFFS_PATH = path.join(__dirname, 'CHRONOLOGICAL_DIFF_KEYS.json');
-
 /*
+
 NOTE to devs: In order to test this file locally, you must:
 1. Have an AWS account with our organization with sufficient permissions set up on your local machine.
 2. You must manually load these AWS credentials into your shell before launching 'docker-compose up'.
@@ -37,19 +33,15 @@ node run_unverified_test_center_diff_merges.js
 It is not recommended to run this code locally unless you are developing a new feature that requires it.
 */
 
-async function checkAWSAccount(callback){
-    const sts = new AWS.STS();
-    const ident = await sts.getCallerIdentity({}).promise();
-    return ident;
-}
+const path = require('path');
+const MODELS_DIR = path.join(__dirname, '../server/db', 'models')
+const db = require(MODELS_DIR)
+const { Op } = require('sequelize')
+const awsUtils = require('./node_helpers/awsUtils')
+const fileUtils = require('./node_helpers/fileUtils')
 
-async function loadDiffFile(){
-    const file = await fs.readFile(S3_UNVER_DIFFS_PATH);
-    const jsonData = JSON.parse(file);
-    return jsonData;
-}
-
-// NOTE: Order of CHRONOLOGICAL_S3_DIFF_KEYS array must be preserved.
+// Order of CHRONOLOGICAL_S3_DIFF_KEYS array must be preserved.
+// If keys are run out-of-order, different results may be produced in our database.
 async function identifyNewDiffs(diffObj, priorDiffs){
     let pendingDiffKeys = diffObj.CHRONOLOGICAL_S3_DIFF_KEYS;
     let newDiffKeys = [];
@@ -72,16 +64,6 @@ async function loadPriorDiffs(){
         return diff.diff_s3_url
     });
     return mappedDiffRecords;
-}
-
-async function loadNewDiffFromS3(diffS3Key){
-    // load object from S3
-    const params = {
-        Bucket: 'staging-gtc-data-batches',
-        Key: 'unver-staged-jobs/' + diffS3Key
-    }
-    const diff = await s3.getObject(params).promise();
-    return diff;
 }
 
 async function runDiffInstallationTransaction(unverDiffKey, diffObj){
@@ -134,14 +116,14 @@ async function insertUnverDiff(unverDiffKey, transaction){
     return unverDiff;
 }
 
-// NOTE: Sequence of diff application is crucial. Order of original list must be preserved.
+// Sequence of diff application is important. Order of original list must be preserved.
 // If a failure occurs, halt in place.
 async function handleAllNewDiffsSequentially(newDiffKeysArr){
     for(let i = 0; i < newDiffKeysArr.length; i++){
         const unverDiffKey = newDiffKeysArr[i];
 
         try{
-            const diffObj = await loadNewDiffFromS3(unverDiffKey);
+            const diffObj = await awsUtils.loadNewDiffFromS3(unverDiffKey);
             const diffBody = JSON.parse(diffObj['Body']);
             console.log(JSON.stringify(diffBody, null, '\t'));
 
@@ -155,11 +137,11 @@ async function handleAllNewDiffsSequentially(newDiffKeysArr){
 }
 
 async function checkAndLoadUnverifiedDiffs(){
-    const ident = await checkAWSAccount();
+    const ident = await awsUtils.checkAWSAccount();
     console.log('AWS ident loaded: ', ident);
     console.log('DB MODELS DIR: ', MODELS_DIR);
 
-    let diffFile = await loadDiffFile();
+    let diffFile = await fileUtils.loadDiffFile();
     console.log('diff file: ', diffFile);
 
     let priorDiffs = await loadPriorDiffs();
